+++ Kernel/GenericInterface/Operation/Ticket/TicketCreate.pm
@@ -1143,13 +1143,21 @@ sub _CheckAttachment {
     my $Attachment = $Param{Attachment};
 
     # check attachment item internally
-    for my $Needed (qw(Content ContentType Filename)) {
+    for my $Needed (qw(Content ContentType)) {
         if ( !$Attachment->{$Needed} ) {
             return {
                 ErrorCode    => 'TicketCreate.MissingParameter',
                 ErrorMessage => "TicketCreate: Attachment->$Needed  parameter is missing!",
             };
         }
+    }
+
+    # Check Filename separately because it can be '0'. See bug#14761.
+    if ( !IsStringWithData( $Attachment->{Filename} ) ) {
+        return {
+            ErrorCode    => 'TicketCreate.MissingParameter',
+            ErrorMessage => "TicketCreate: Attachment->Filename  parameter is missing!",
+        };
     }
 
     # check Article->ContentType