+++ Kernel/GenericInterface/Operation/Ticket/TicketUpdate.pm
@@ -1243,13 +1243,21 @@ sub _CheckAttachment {
     }
 
     # check attachment item internally
-    for my $Needed (qw(Content ContentType Filename)) {
+    for my $Needed (qw(Content ContentType)) {
         if ( !$Attachment->{$Needed} ) {
             return {
                 ErrorCode    => 'TicketUpdate.MissingParameter',
                 ErrorMessage => "TicketUpdate: Attachment->$Needed parameter is missing!",
             };
         }
+    }
+
+    # Check Filename separately because it can be '0'. See bug#14761.
+    if ( !IsStringWithData( $Attachment->{Filename} ) ) {
+        return {
+            ErrorCode    => 'TicketCreate.MissingParameter',
+            ErrorMessage => "TicketCreate: Attachment->Filename  parameter is missing!",
+        };
     }
 
     # check Article->ContentType